resolves conflicts with previous merge. jService API is now called from GameBoardView.js
