--- src/components/App/App.js
@@ -1,36 +1,0 @@
-import React, { useEffect } from 'react';
-import jService from '../../api/jService';
-import filterCluesWithValues from '../../logic/getCategoryClues';
-import generateCategoryIdSet from '../../logic/generateCategoryIdSet';
-
-function App() {
-  useEffect(() => {
-    getClues();
-  }, []);
-
-  const getClues = async (categoryID) => {
-    const { data } = await jService.get('/', {
-      params: {
-        id: ids[0],
-      },
-    });
-    // getClues will need to be called for each category column.
-    // It will need a categoryID which is generated by generateCategoryIdSet().
-    return console.log(filterCluesWithValues(data.clues, data.title));
-  };
-
-  const ids = generateCategoryIdSet();
-
-  const catergoryIds = ids.map((id) => {
-    return <p key={id}>{id}</p>;
-  });
-
-  return (
-    <div className="App">
-      <h1>Welcome to the Jeopardy Application</h1>
-      {catergoryIds}
-    </div>
-  );
-}
-
-export default App;
