+++ src/components/views/GameBoardView/GameBoardView.js
@@ -1,33 +1,56 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import GameBoardColumn from './GameBoard_components/GameBoardColumn';
-import GameCard from './GameBoard_components/GameCard'
+import GameCard from './GameBoard_components/GameCard';
+
+import jService from '../../../api/jService';
+import filterCluesWithValues from '../../../logic/getCategoryClues';
+import generateCategoryIdSet from '../../../logic/generateCategoryIdSet';
 
 const GameBoardView = () => {
-const Cards =
-            [
-                <GameCard Category="dogs" />, 
-                <GameCard 
-                    Question="This dog is spotted and often referred to in movie containing the number 101"
-                    PointValue="100"
-                    Answer="Dalmation"
-                />, 
-                <GameCard />, 
-                <GameCard />, 
-                <GameCard />, 
-                <GameCard />
-            ];
-    return (
-        <div className="flex justify-between items-center w-full space-x-4">
-            {/* <h1>Game Board Goes Here</h1> */}
-            {/* <GameCard /> */}
-            <GameBoardColumn Cards={Cards} />
-            <GameBoardColumn Cards={Cards} />
-            <GameBoardColumn Cards={Cards} />
-            <GameBoardColumn Cards={Cards} />
-            <GameBoardColumn Cards={Cards} />
-            <GameBoardColumn Cards={Cards} />
-        </div>
-    )
-}
+  // each GameCard component should recieve getClues as a prop.
+  // For now I'm just calling getClues from here for testing.
+  useEffect(() => {
+    getClues();
+  }, []);
+
+  const ids = generateCategoryIdSet();
+
+  const getClues = async (categoryID) => {
+    const { data } = await jService.get('/', {
+      params: {
+        // this will need to be change to categoryID
+        id: ids[0],
+      },
+    });
+    // getClues will need to be called for each category column.
+    // It will need a categoryID which is generated by generateCategoryIdSet().
+    return console.log(filterCluesWithValues(data.clues, data.title));
+  };
+
+  const Cards = [
+    <GameCard Category="dogs" />,
+    <GameCard
+      Question="This dog is spotted and often referred to in movie containing the number 101"
+      PointValue="100"
+      Answer="Dalmation"
+    />,
+    <GameCard />,
+    <GameCard />,
+    <GameCard />,
+    <GameCard />,
+  ];
+  return (
+    <div className="flex justify-between items-center w-full space-x-4">
+      {/* <h1>Game Board Goes Here</h1> */}
+      {/* <GameCard /> */}
+      <GameBoardColumn Cards={Cards} getClues={getClues} />
+      <GameBoardColumn Cards={Cards} getClues={getClues} />
+      <GameBoardColumn Cards={Cards} getClues={getClues} />
+      <GameBoardColumn Cards={Cards} getClues={getClues} />
+      <GameBoardColumn Cards={Cards} getClues={getClues} />
+      <GameBoardColumn Cards={Cards} getClues={getClues} />
+    </div>
+  );
+};
 
 export default GameBoardView;
